+++ CHANGELOG.md
@@ -12,7 +12,35 @@
 
 - ([#PR](link))
 
-# [v0.3.6](https://github.com/pybamm-team/liionpack/)
+# [v0.3.9](https://github.com/pybamm-team/liionpack/tree/v0.3.9) - 2024-04-26
+
+## Features
+
+- Update PyBaMM to 24.1 ([#293](https://github.com/pybamm-team/liionpack/pull/293))
+
+## Bug fixes
+
+- Fix event reporting ([#288](https://github.com/pybamm-team/liionpack/pull/288))
+- Fix experiments with small time steps ([#289](https://github.com/pybamm-team/liionpack/pull/289))
+- Remove the Jax install from CI ([#287](https://github.com/pybamm-team/liionpack/pull/287))
+- ([#PR](link))
+
+
+# [v0.3.8](https://github.com/pybamm-team/liionpack/tree/v0.3.8) - 2024-01-20
+
+## Bug fixes
+
+- Update to PyBaMM 23.9
+
+# [v0.3.7](https://github.com/pybamm-team/liionpack/tree/v0.3.7) - 2023-07-05
+
+
+## Bug fixes
+
+- Update to PyBaMM 23.5 with small chages to protocols ([#265](https://github.com/pybamm-team/liionpack/pull/265))
+
+
+# [v0.3.6](https://github.com/pybamm-team/liionpack/tree/v0.3.6) - 2023-05-26
 
 
 ## Bug fixes
@@ -20,7 +48,7 @@
 - Fix a RunTimeError introduced by change to latest version of casadi that PyBaMM now supports. Solution returned by casadi no longer contains initial state. ([#259](https://github.com/pybamm-team/liionpack/pull/259))
 
 
-# [v0.3.5](https://github.com/pybamm-team/liionpack/) - 2023-04-05
+# [v0.3.5](https://github.com/pybamm-team/liionpack/tree/v0.3.5) - 2023-04-05
 
 ## Features
 
@@ -39,7 +67,7 @@
 
 - Update pre-commit hooks ([#247](https://github.com/pybamm-team/liionpack/pull/247))
 
-# [v.0.3.4](https://github.com/pybamm-team/liionpack/tree/v.0.3.4) - 2023-03-03
+# [v0.3.4](https://github.com/pybamm-team/liionpack/tree/v.0.3.4) - 2023-03-03
 
 ## This version was yanked from PyPi due to an accidental early release that addressed changes coming into PyBaMM in version 23.3.